fix runner/exec
